updated interview schedule delete and recruitment delete update

+++ app/Services/InterviewScheduleService.php
@@ -10,6 +10,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
+use Illuminate\Validation\ValidationException;
 
 
 class InterviewScheduleService
@@ -36,7 +37,7 @@ class InterviewScheduleService
     }
 
 
-    public function getSchedulesByStepId(int $id):mixed
+    public function getSchedulesByStepId(int $id): mixed
     {
         $scheduleBuilder = InterviewSchedule::select([
             'interview_schedules.id',
@@ -83,7 +84,23 @@ class InterviewScheduleService
      */
     public function destroy(InterviewSchedule $schedule): bool
     {
-        return $schedule->delete();
+        $scheduledCandidates = $this->countCandidatePerScheduled($schedule->id);
+
+        if ($scheduledCandidates == 0) {
+            return $schedule->delete();
+        } else {
+            return false;
+        }
+    }
+
+    /**
+     * @param int $scheduleId
+     * @return mixed
+     */
+    public function countCandidatePerScheduled(int $scheduleId): mixed
+    {
+        return CandidateInterview::where('interview_schedule_id', $scheduleId)->count('id');
+
     }
 
 
@@ -133,7 +150,7 @@ class InterviewScheduleService
         if (!empty($request['applied_job_ids'])) {
             $request['applied_job_ids'] = isset($request['applied_job_ids']) && is_array($request['applied_job_ids']) ? $request['applied_job_ids'] : explode(',', $request['applied_job_ids']);
         }
-        Log::info("------->>", $request['applied_job_ids']);
+
         $rules = [
             'notify' => [
                 'required',
@@ -225,7 +242,7 @@ class InterviewScheduleService
 
         foreach ($appliedJobIds as $appliedJobId) {
             $appliedJob = AppliedJob::findOrFail($appliedJobId);
-           CandidateInterview::where('applied_job_id', $appliedJob->id)
+            CandidateInterview::where('applied_job_id', $appliedJob->id)
                 ->where('recruitment_step_id', $appliedJob->current_recruitment_step_id)
                 ->where('interview_schedule_id', $scheduleId)
                 ->delete();
